+++ spec/ruby/core/array/rindex_spec.rb
@@ -32,6 +32,22 @@ describe "Array#rindex" do
     [1, 1, 3, 2, 1, 3].rindex(4).should == nil
   end
 
+  it "does not search before a non-negative offset" do
+    [1, 2, 1, 2].rindex(1, offset: 1).should == 2
+    [1, 2, 1, 2].rindex(1, offset: 3).should == nil
+  end
+
+  it "starts searching at a negative offset relative to the end" do
+    [1, 2, 1, 2].rindex(2, offset: -2).should == 1
+  end
+
+  it "returns nil when the offset is outside the array" do
+    array = [1, 2, 1, 2]
+
+    array.rindex(1, offset: array.size).should == nil
+    array.rindex(1, offset: -array.size - 1).should == nil
+  end
+
   it "returns correct index even after delete_at" do
     array = ["fish", "bird", "lion", "cat"]
     array.delete_at(0)
@@ -57,6 +73,12 @@ describe "Array#rindex" do
   it "ignores the block if there is an argument" do
     -> {
       [4, 2, 1, 5, 1, 3].rindex(5) { |x| x < 2 }.should == 3
+    }.should complain(/given block not used/)
+  end
+
+  it "ignores the block if there is an argument and an offset" do
+    -> {
+      [4, 2, 1, 5, 1, 3].rindex(1, offset: 2) { |x| x < 2 }.should == 4
     }.should complain(/given block not used/)
   end